refactor: Complete Phase 1 & 4 - Debug consolidation and utility streamlining
Phase 1 - Debug Infrastructure Cleanup:
- Consolidate debug tools into unified src/debug/debug_manager.py
- Remove duplicate debug_window_structure.py and debug_helper.py
- Support both interactive and programmatic debug modes
- Update CLAUDE.md with new debug interface documentation

Phase 4 - Utility Consolidation:
- Merge id_generator.py + string_processor.py → text_utils.py
- Integrate input_validator.py → ui/user_interaction.py
- Move data_loader.py functionality → main.py
- Update all import references to new consolidated modules
- Remove 6 redundant utility files

Results:
- Total files eliminated: 6 utility + 2 debug files = 8 files
- Utils directory: 6 → 2 files (67% reduction)
- All functionality preserved and properly organized
- No circular dependencies, cleaner import structure
