+++ src/debug/__init__.py
@@ -1,0 +1,17 @@
+"""Debug module for RPA window structure analysis."""
+
+from .debug_manager import (
+    UnifiedDebugManager,
+    create_debug_manager,
+    quick_window_analysis,
+    monitor_circulation_dialogs,
+    run_interactive_debug
+)
+
+__all__ = [
+    'UnifiedDebugManager',
+    'create_debug_manager', 
+    'quick_window_analysis',
+    'monitor_circulation_dialogs',
+    'run_interactive_debug'
+]
