+++ src/services/task_card_service.py
@@ -4,7 +4,7 @@ from typing import List, Optional
 
 from services.dialog_service import DialogHandler
 from services.chroma_service import ChromaService
-from utils.input_validator import SelectionResult
+from ui.user_interaction import SelectionResult
 from utils.error_handler import setup_logger
 
 logger = setup_logger(__name__)
